Support PowerStream Messages with Multiple Payloads (#95)
* Supper PowerStream Messages with Multiple Payloads

When the mobile app is launched a request is sent out to request all the data. This change allows Home Assistant to read extra messages that ProtoBuf doesn't read naturally.

* Add Beta flag to PowerStream in ReadMe

To help keep requests and miss understandings to a minimum I have added notes to the PowerStream section of the readme with the limitations of the current implementation.